+++ orchestrator/__init__.py
@@ -1,2 +1,2 @@
-"""orchestrator service v0.4.0"""
-__version__ = "0.4.0"
+"""orchestrator service v0.5.0"""
+__version__ = "0.5.0"